requirements.txt
argparse
fpdf
Pillow